+++ system_tests/noxfile.py
@@ -30,7 +30,7 @@ import nox
 import py.path
 
 HERE = os.path.abspath(os.path.dirname(__file__))
-LIBRARY_DIR = os.path.join(HERE, "..")
+LIBRARY_DIR = os.path.abspath(os.path.dirname(HERE))
 DATA_DIR = os.path.join(HERE, "data")
 SERVICE_ACCOUNT_FILE = os.path.join(DATA_DIR, "service_account.json")
 AUTHORIZED_USER_FILE = os.path.join(DATA_DIR, "authorized_user.json")
@@ -169,16 +169,26 @@ def configure_cloud_sdk(session, application_default_credentials, project=False)
 # Test sesssions
 
 TEST_DEPENDENCIES_ASYNC = ["aiohttp", "pytest-asyncio", "nest-asyncio"]
-TEST_DEPENDENCIES_SYNC = ["pytest", "requests"]
+TEST_DEPENDENCIES_SYNC = ["pytest", "requests", "mock"]
 PYTHON_VERSIONS_ASYNC = ["3.7"]
 PYTHON_VERSIONS_SYNC = ["2.7", "3.7"]
+
+
+def default(session, *test_paths):
+    # replace 'session._runner.friendly_name' with
+    # session.name once nox has released a new version
+    # https://github.com/theacodes/nox/pull/386
+    sponge_log = f"--junitxml=system_{str(session._runner.friendly_name)}_sponge_log.xml"
+    session.run(
+        "pytest", sponge_log, *test_paths,
+    )
 
 
 @nox.session(python=PYTHON_VERSIONS_SYNC)
 def service_account_sync(session):
     session.install(*TEST_DEPENDENCIES_SYNC)
     session.install(LIBRARY_DIR)
-    session.run("pytest", "system_tests_sync/test_service_account.py")
+    default(session, "system_tests_sync/test_service_account.py")
 
 
 @nox.session(python=PYTHON_VERSIONS_SYNC)
@@ -187,7 +197,11 @@ def default_explicit_service_account(session):
     session.env[EXPECT_PROJECT_ENV] = "1"
     session.install(*TEST_DEPENDENCIES_SYNC)
     session.install(LIBRARY_DIR)
-    session.run("pytest", "system_tests_sync/test_default.py", "system_tests_sync/test_id_token.py")
+    default(
+        session,
+        "system_tests_sync/test_default.py",
+        "system_tests_sync/test_id_token.py",
+    )
 
 
 @nox.session(python=PYTHON_VERSIONS_SYNC)
@@ -195,7 +209,9 @@ def default_explicit_authorized_user(session):
     session.env[EXPLICIT_CREDENTIALS_ENV] = AUTHORIZED_USER_FILE
     session.install(*TEST_DEPENDENCIES_SYNC)
     session.install(LIBRARY_DIR)
-    session.run("pytest", "system_tests_sync/test_default.py")
+    default(
+        session, "system_tests_sync/test_default.py",
+    )
 
 
 @nox.session(python=PYTHON_VERSIONS_SYNC)
@@ -205,7 +221,9 @@ def default_explicit_authorized_user_explicit_project(session):
     session.env[EXPECT_PROJECT_ENV] = "1"
     session.install(*TEST_DEPENDENCIES_SYNC)
     session.install(LIBRARY_DIR)
-    session.run("pytest", "system_tests_sync/test_default.py")
+    default(
+        session, "system_tests_sync/test_default.py",
+    )
 
 
 @nox.session(python=PYTHON_VERSIONS_SYNC)
@@ -214,7 +232,9 @@ def default_cloud_sdk_service_account(session):
     session.env[EXPECT_PROJECT_ENV] = "1"
     session.install(*TEST_DEPENDENCIES_SYNC)
     session.install(LIBRARY_DIR)
-    session.run("pytest", "system_tests_sync/test_default.py")
+    default(
+        session, "system_tests_sync/test_default.py",
+    )
 
 
 @nox.session(python=PYTHON_VERSIONS_SYNC)
@@ -222,7 +242,9 @@ def default_cloud_sdk_authorized_user(session):
     configure_cloud_sdk(session, AUTHORIZED_USER_FILE)
     session.install(*TEST_DEPENDENCIES_SYNC)
     session.install(LIBRARY_DIR)
-    session.run("pytest", "system_tests_sync/test_default.py")
+    default(
+        session, "system_tests_sync/test_default.py",
+    )
 
 
 @nox.session(python=PYTHON_VERSIONS_SYNC)
@@ -231,7 +253,10 @@ def default_cloud_sdk_authorized_user_configured_project(session):
     session.env[EXPECT_PROJECT_ENV] = "1"
     session.install(*TEST_DEPENDENCIES_SYNC)
     session.install(LIBRARY_DIR)
-    session.run("pytest", "system_tests_sync/test_default.py")
+    default(
+        session, "system_tests_sync/test_default.py",
+    )
+
 
 @nox.session(python=PYTHON_VERSIONS_SYNC)
 def compute_engine(session):
@@ -240,7 +265,9 @@ def compute_engine(session):
     # credentials are detected from environment
     del session.virtualenv.env["GOOGLE_APPLICATION_CREDENTIALS"]
     session.install(LIBRARY_DIR)
-    session.run("pytest", "system_tests_sync/test_compute_engine.py")
+    default(
+        session, "system_tests_sync/test_compute_engine.py",
+    )
 
 
 @nox.session(python=["2.7"])
@@ -249,6 +276,7 @@ def app_engine(session):
         session.log("Skipping App Engine tests.")
         return
 
+    session.install(LIBRARY_DIR)
     # Unlike the default tests above, the App Engine system test require a
     # 'real' gcloud sdk installation that is configured to deploy to an
     # app engine project.
@@ -269,9 +297,8 @@ def app_engine(session):
     application_url = GAE_APP_URL_TMPL.format(GAE_TEST_APP_SERVICE, project_id)
 
     # Vendor in the test application's dependencies
-    session.chdir(os.path.join(HERE, "../app_engine_test_app"))
+    session.chdir(os.path.join(HERE, "system_tests_sync/app_engine_test_app"))
     session.install(*TEST_DEPENDENCIES_SYNC)
-    session.install(LIBRARY_DIR)
     session.run(
         "pip", "install", "--target", "lib", "-r", "requirements.txt", silent=True
     )
@@ -282,15 +309,39 @@ def app_engine(session):
     # Run the tests
     session.env["TEST_APP_URL"] = application_url
     session.chdir(HERE)
-    session.run("pytest", "system_tests_sync/test_app_engine.py")
+    default(
+        session, "system_tests_sync/test_app_engine.py",
+    )
 
 
 @nox.session(python=PYTHON_VERSIONS_SYNC)
 def grpc(session):
     session.install(LIBRARY_DIR)
-    session.install(*TEST_DEPENDENCIES_SYNC, "google-cloud-pubsub==1.0.0")
+    session.install(*TEST_DEPENDENCIES_SYNC, "google-cloud-pubsub==1.7.0")
     session.env[EXPLICIT_CREDENTIALS_ENV] = SERVICE_ACCOUNT_FILE
-    session.run("pytest", "system_tests_sync/test_grpc.py")
+    default(
+        session, "system_tests_sync/test_grpc.py",
+    )
+
+
+@nox.session(python=PYTHON_VERSIONS_SYNC)
+def requests(session):
+    session.install(LIBRARY_DIR)
+    session.install(*TEST_DEPENDENCIES_SYNC)
+    session.env[EXPLICIT_CREDENTIALS_ENV] = SERVICE_ACCOUNT_FILE
+    default(
+        session, "system_tests_sync/test_requests.py",
+    )
+
+
+@nox.session(python=PYTHON_VERSIONS_SYNC)
+def urllib3(session):
+    session.install(LIBRARY_DIR)
+    session.install(*TEST_DEPENDENCIES_SYNC)
+    session.env[EXPLICIT_CREDENTIALS_ENV] = SERVICE_ACCOUNT_FILE
+    default(
+        session, "system_tests_sync/test_urllib3.py",
+    )
 
 
 @nox.session(python=PYTHON_VERSIONS_SYNC)
@@ -298,15 +349,26 @@ def mtls_http(session):
     session.install(LIBRARY_DIR)
     session.install(*TEST_DEPENDENCIES_SYNC, "pyopenssl")
     session.env[EXPLICIT_CREDENTIALS_ENV] = SERVICE_ACCOUNT_FILE
-    session.run("pytest", "system_tests_sync/test_mtls_http.py")
+    default(
+        session, "system_tests_sync/test_mtls_http.py",
+    )
 
-#ASYNC SYSTEM TESTS
+
+@nox.session(python=PYTHON_VERSIONS_SYNC)
+def external_accounts(session):
+    session.install(*TEST_DEPENDENCIES_SYNC, "google-auth", "google-api-python-client", "enum34")
+    default(session, "system_tests_sync/test_external_accounts.py")
+
+
+# ASYNC SYSTEM TESTS
 
 @nox.session(python=PYTHON_VERSIONS_ASYNC)
 def service_account_async(session):
-    session.install(*(TEST_DEPENDENCIES_SYNC+TEST_DEPENDENCIES_ASYNC))
+    session.install(*(TEST_DEPENDENCIES_SYNC + TEST_DEPENDENCIES_ASYNC))
     session.install(LIBRARY_DIR)
-    session.run("pytest", "system_tests_async/test_service_account.py")
+    default(
+        session, "system_tests_async/test_service_account.py",
+    )
 
 
 @nox.session(python=PYTHON_VERSIONS_ASYNC)
@@ -315,8 +377,11 @@ def default_explicit_service_account_async(session):
     session.env[EXPECT_PROJECT_ENV] = "1"
     session.install(*(TEST_DEPENDENCIES_SYNC + TEST_DEPENDENCIES_ASYNC))
     session.install(LIBRARY_DIR)
-    session.run("pytest", "system_tests_async/test_default.py", 
-    "system_tests_async/test_id_token.py")
+    default(
+        session,
+        "system_tests_async/test_default.py",
+        "system_tests_async/test_id_token.py",
+    )
 
 
 @nox.session(python=PYTHON_VERSIONS_ASYNC)
@@ -324,7 +389,9 @@ def default_explicit_authorized_user_async(session):
     session.env[EXPLICIT_CREDENTIALS_ENV] = AUTHORIZED_USER_FILE
     session.install(*(TEST_DEPENDENCIES_SYNC + TEST_DEPENDENCIES_ASYNC))
     session.install(LIBRARY_DIR)
-    session.run("pytest", "system_tests_async/test_default.py")
+    default(
+        session, "system_tests_async/test_default.py",
+    )
 
 
 @nox.session(python=PYTHON_VERSIONS_ASYNC)
@@ -334,7 +401,9 @@ def default_explicit_authorized_user_explicit_project_async(session):
     session.env[EXPECT_PROJECT_ENV] = "1"
     session.install(*(TEST_DEPENDENCIES_SYNC + TEST_DEPENDENCIES_ASYNC))
     session.install(LIBRARY_DIR)
-    session.run("pytest", "system_tests_async/test_default.py")
+    default(
+        session, "system_tests_async/test_default.py",
+    )
 
 
 @nox.session(python=PYTHON_VERSIONS_ASYNC)
@@ -343,7 +412,9 @@ def default_cloud_sdk_service_account_async(session):
     session.env[EXPECT_PROJECT_ENV] = "1"
     session.install(*(TEST_DEPENDENCIES_SYNC + TEST_DEPENDENCIES_ASYNC))
     session.install(LIBRARY_DIR)
-    session.run("pytest", "system_tests_async/test_default.py")
+    default(
+        session, "system_tests_async/test_default.py",
+    )
 
 
 @nox.session(python=PYTHON_VERSIONS_ASYNC)
@@ -351,7 +422,9 @@ def default_cloud_sdk_authorized_user_async(session):
     configure_cloud_sdk(session, AUTHORIZED_USER_FILE)
     session.install(*(TEST_DEPENDENCIES_SYNC + TEST_DEPENDENCIES_ASYNC))
     session.install(LIBRARY_DIR)
-    session.run("pytest", "system_tests_async/test_default.py")
+    default(
+        session, "system_tests_async/test_default.py",
+    )
 
 
 @nox.session(python=PYTHON_VERSIONS_ASYNC)
@@ -360,4 +433,6 @@ def default_cloud_sdk_authorized_user_configured_project_async(session):
     session.env[EXPECT_PROJECT_ENV] = "1"
     session.install(*(TEST_DEPENDENCIES_SYNC + TEST_DEPENDENCIES_ASYNC))
     session.install(LIBRARY_DIR)
-    session.run("pytest", "system_tests_async/test_default.py")
+    default(
+        session, "system_tests_async/test_default.py",
+    )